feat(lint): gate CFG-based lint behind lint-v2 feature (keep v1 as default)
