+++ crates/tinymist-query/Cargo.toml
@@ -71,6 +71,7 @@ hex = { version = "0.4" }
 [features]
 local-registry = ["tinymist-world/system"]
 system = ["tinymist-project/system"]
+lint-v2 = ["tinymist-lint/lint-v2"]
 
 [lints]
 workspace = true
