+++ crates/tinymist/src/project.rs
@@ -29,8 +29,12 @@ use reflexo::hash::FxHashMap;
 use sync_ls::{LspClient, TypedLspClient};
 use tinymist_project::vfs::{FileChangeSet, MemoryEvent};
 use tinymist_query::analysis::{Analysis, LspQuerySnapshot, PeriscopeProvider};
+#[cfg(not(feature = "lint-v2"))]
+use tinymist_query::CheckRequest;
+#[cfg(feature = "lint-v2")]
+use tinymist_query::LintRequest;
 use tinymist_query::{
-    CompilerQueryRequest, DiagnosticsMap, LintRequest, LocalContext, SemanticRequest, GLOBAL_STATS,
+    CompilerQueryRequest, DiagnosticsMap, LocalContext, SemanticRequest, GLOBAL_STATS,
 };
 use tinymist_render::PeriscopeRenderer;
 use tinymist_std::{error::prelude::*, ImmutPath};
@@ -559,15 +563,39 @@ impl LintHook {
             let mut ctx = analysis.enter(snap.graph.clone());
 
             // todo: check all errors in this file
-            let Some(lint) = LintRequest { snap }.request(&mut ctx) else {
-                return;
-            };
+            #[cfg(feature = "lint-v2")]
+            {
+                let Some(lint) = LintRequest { snap }.request(&mut ctx) else {
+                    return;
+                };
 
-            log::trace!("notify lint diagnostics({:?}): {:#?}", dv.id, lint);
+                log::trace!("notify lint diagnostics({:?}): {:#?}", dv.id, lint);
 
-            editor_tx
-                .send(EditorRequest::Diag(dv, DiagKind::Lint, Some(lint)))
-                .log_error("failed to send lint diagnostics");
+                editor_tx
+                    .send(EditorRequest::Diag(dv, DiagKind::Lint, Some(lint)))
+                    .log_error("failed to send lint diagnostics");
+            }
+
+            #[cfg(not(feature = "lint-v2"))]
+            {
+                let Some(diagnostics) = CheckRequest { snap }.request(&mut ctx) else {
+                    return;
+                };
+
+                log::trace!(
+                    "notify lint diagnostics({:?}): {:#?}",
+                    dv.id,
+                    diagnostics.lint
+                );
+
+                editor_tx
+                    .send(EditorRequest::Diag(
+                        dv,
+                        DiagKind::Lint,
+                        Some(diagnostics.lint),
+                    ))
+                    .log_error("failed to send lint diagnostics");
+            }
         });
     }
 }
